+++ derive_builder_test/tests/compiletests.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "compiletests")]
+#![cfg(feature = "nightlytests")]
 extern crate compiletest_rs as compiletest;
 
 // note: